refactor(web): group components into chrome/ and summaries/
Move the flat app-shell components (container, header, footer, nav,
layout, theme-toggle) under components/chrome/, and the shared
presentational summary fragments (character-summary, weapon-summary)
under components/summaries/. Relocate artifact-planner into
features/teams/ since it is imported only by the team member planner.

Issue #599 originally proposed components/teams/ and
components/characters/ directories, but the app has since adopted the
features/ architecture (features/teams/, features/collection/), which
already organized the team-building components it named. This completes
the remaining flat components/ directory in line with that convention.

Closes #599
