+++ apps/web/src/components/chrome/footer.test.tsx
