+++ apps/web/src/components/summaries/character-summary.tsx
