+++ apps/web/src/components/summaries/weapon-summary.tsx
