+++ apps/web/src/features/collection/weapons/weapon-card.tsx
@@ -4,7 +4,7 @@
 import type { Weapon } from '@genshin/game-data';
 import type { JSX } from 'react';
 
-import { WeaponSummary } from '@/components/weapon-summary';
+import { WeaponSummary } from '@/components/summaries/weapon-summary';
 import { RARITY_BORDER_COLORS } from '@/lib/rarity-styles';
 import { cn } from '@/lib/utils';
