+++ apps/web/src/pages/teams-page.tsx
@@ -7,7 +7,7 @@ import { getCharacterById } from '@genshin/game-data';
 import type { JSX } from 'react';
 import { useCallback, useMemo, useState } from 'react';
 
-import { Container } from '@/components/container';
+import { Container } from '@/components/chrome/container';
 import { Sheet, SheetContent, SheetHeader } from '@/components/ui/sheet';
 import { useCollection } from '@/features/collection/characters/use-character-collection';
 import { useWeaponCollection } from '@/features/collection/weapons/use-weapon-collection';
